task5 -> refactored routes
