+++ package.json
@@ -26,7 +26,9 @@
     "eventemitter3": "2.0.3",
     "express": "4.16.2",
     "glob": "7.1.2",
+    "jsonwebtoken": "8.1.0",
     "minimist": "1.2.0",
+    "passport": "0.4.0",
     "request": "2.83.0",
     "split2": "2.2.0",
     "through2": "2.0.3"
